Change handling of hints for preventing the transliteration
If the text in (X)HTML is inside <span lang="sr-Latn"></span> or <span lang="sr-Cyrl"></span> it won't be transliterated to other script.

Plain text between <| and |> won't be transliterated from latin to cyrillic. This is still not implemented the other way round.

Also, the lang attribute of the html element is set in a way that properly describes the transformation that is performed on the text: "sr-Cyrl-t-sr-Latn" if a latin text is tranliterated to the cyrilic script and "sr-Latn-t-sr-Cyrl" if a cyrilic text is transiterated to the latin script.